First draft of LegacyCapitalGainsCalculator using streams

+++ src/main/java/com/example/challenges/finance/domain/TaxReportItem.java
@@ -1,3 +1,5 @@
 package com.example.challenges.finance.domain;
 
-public record TaxReportItem(String ticker, Double totalTax) {}
+import java.math.BigDecimal;
+
+public record TaxReportItem(String ticker, BigDecimal totalTax) {}
